cohorts/2023/week_2_workflow_orchestration/requirements.txt
pandas
prefect
prefect-gcp[cloud_storage]
python-dotenv
pyarrow
pandas-gbq
